infra/pyproject.toml
[project]
name = "inspect-ec2-sandbox-infra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aws-cdk-lib>=2.209.1",
]
